+++ src/exchange/mod.rs
@@ -3,3 +3,5 @@
 pub mod binance;
 pub mod coinbase;
 pub mod nash;
+pub mod traits;
+pub mod any;
