+++ src/prelude/mod.rs
@@ -1,7 +1,7 @@
 //! This module contains some Traits and Structs that are frequently used.
 pub use crate::model::{OrderBookRequest, OrderBookResponse};
 
-pub use crate::exchange_ws::ExchangeWs;
-pub use crate::exchange_info::{ExchangeInfo, ExchangeInfoRetrieval, MarketPair, MarketPairHandle};
-pub use crate::exchange_traits::{Exchange, ExchangeAccount, ExchangeMarketData};
+pub use crate::exchange::traits::stream::ExchangeWs;
+pub use crate::exchange::traits::info::{ExchangeInfo, ExchangeInfoRetrieval, MarketPair, MarketPairHandle};
+pub use crate::exchange::traits::{Exchange, ExchangeAccount, ExchangeMarketData};
