+++ tests/mod.rs
@@ -1,6 +1,6 @@
 extern crate openlimits;
 
-mod any_exchange;
+mod exchange;
 mod apis;
 mod binance;
 mod coinbase;
