+++ tests/nash/websocket.rs
@@ -1,17 +1,19 @@
-use dotenv::dotenv;
-use nash_native_client::Environment;
-use openlimits::OpenLimits;
-use openlimits::prelude::*;
-use openlimits::exchange_ws::ExchangeWs;
-use openlimits::model::websocket::AccountOrders;
-use openlimits::model::{CancelAllOrdersRequest, OpenLimitOrderRequest, TimeInForce};
-use openlimits::exchange::nash::{Nash, NashCredentials, NashParameters};
-use openlimits::{exchange_ws::OpenLimitsWs, model::websocket::Subscription, exchange::nash::NashWebsocket};
-use rust_decimal::Decimal;
+use std::{env, sync::mpsc::sync_channel};
 use std::str::FromStr;
 use std::time::Duration as NativeDuration;
-use std::{env, sync::mpsc::sync_channel};
+
+use dotenv::dotenv;
+use nash_native_client::Environment;
+use rust_decimal::Decimal;
 use tokio::time::Duration;
+
+use openlimits::{exchange::nash::NashWebsocket, model::websocket::Subscription};
+use openlimits::exchange::nash::{Nash, NashCredentials, NashParameters};
+use openlimits::exchange::traits::stream::{ExchangeWs, OpenLimitsWs};
+use openlimits::model::{CancelAllOrdersRequest, OpenLimitOrderRequest, TimeInForce};
+use openlimits::model::websocket::AccountOrders;
+use openlimits::OpenLimits;
+use openlimits::prelude::*;
 
 async fn init_exchange() -> Nash {
     dotenv().ok();
